Make some usages generic

+++ PKHeX.Core/MysteryGifts/MysteryGift.cs
@@ -31,7 +31,7 @@ namespace PKHeX.Core
             return result;
         }
 
-        public override bool Empty => Data.IsRangeAll(0, 0, Data.Length);
+        public override bool Empty => Data.IsRangeAll((byte)0, 0, Data.Length);
     }
 
     /// <summary>
